Move configs and fix imports
