+++ config/contentPlugins.js
@@ -1,4 +1,4 @@
-const { generatePluginConfig } = require('./src/utils/pluginConfigGenerators');
+const { generatePluginConfig } = require('../src/utils/pluginConfigGenerators');
 const {
   buildPluginsConfig,
   maintainPluginsConfig,
@@ -13,24 +13,24 @@ module.exports = async () => {
     {
       id: 'build',
       globalSidebars: ['build'],
-      path: path.resolve(__dirname, 'docs/build/getting-started'),
+      path: path.resolve(__dirname, '../docs/build/getting-started'),
       sidebarPath: path.resolve(
         __dirname,
-        'docs/build/getting-started/sidebars.ts',
+        '../docs/build/getting-started/sidebars.ts',
       ),
       routeBasePath: 'build',
     },
-    ...generatePluginConfig(buildPluginsConfig, __dirname + '/docs/build/'),
+    ...generatePluginConfig(buildPluginsConfig, path.resolve(__dirname, '../docs/build/')),
     {
       id: 'introduction-docs-stardust',
       path: path.resolve(
         __dirname,
-        'docs/build/introduction-docs/stardust/docs',
+        '../docs/build/introduction-docs/stardust/docs',
       ),
       routeBasePath: 'introduction',
       sidebarPath: path.resolve(
         __dirname,
-        'docs/build/introduction-docs/stardust/sidebars.js',
+        '../docs/build/introduction-docs/stardust/sidebars.js',
       ),
       versions: {
         current: {
@@ -45,15 +45,15 @@ module.exports = async () => {
   const maintainPlugins = [
     ...generatePluginConfig(
       maintainPluginsConfig,
-      __dirname + '/docs/maintain/',
+      path.resolve(__dirname, '../docs/maintain/'),
     ),
     {
       id: 'maintain',
       globalSidebars: ['maintain'],
-      path: path.resolve(__dirname, 'docs/maintain/getting-started'),
+      path: path.resolve(__dirname, '../docs/maintain/getting-started'),
       sidebarPath: path.resolve(
         __dirname,
-        'docs/maintain/getting-started/sidebars.ts',
+        '../docs/maintain/getting-started/sidebars.ts',
       ),
       routeBasePath: 'maintain',
     },
@@ -62,8 +62,8 @@ module.exports = async () => {
   const getStartedPlugins = [
     {
       id: 'get-started',
-      path: path.resolve(__dirname, 'docs/get-started'),
-      sidebarPath: path.resolve(__dirname, 'docs/get-started/sidebars.ts'),
+      path: path.resolve(__dirname, '../docs/get-started/'),
+      sidebarPath: path.resolve(__dirname, '../docs/get-started/sidebars.ts'),
       routeBasePath: 'get-started',
     },
   ];
@@ -71,8 +71,8 @@ module.exports = async () => {
   const learnPlugins = [
     {
       id: 'learn',
-      path: path.resolve(__dirname, 'docs/learn'),
-      sidebarPath: path.resolve(__dirname, 'docs/learn/sidebars.ts'),
+      path: path.resolve(__dirname, '../docs/learn/'),
+      sidebarPath: path.resolve(__dirname, '../docs/learn/sidebars.ts'),
       routeBasePath: 'learn',
     },
   ];
@@ -82,8 +82,8 @@ module.exports = async () => {
     ...getStartedPlugins,
     {
       id: 'community',
-      path: path.resolve(__dirname, 'docs/community'),
-      sidebarPath: path.resolve(__dirname, 'docs/community/sidebars.ts'),
+      path: path.resolve(__dirname, '../docs/community/'),
+      sidebarPath: path.resolve(__dirname, '../docs/community/sidebars.ts'),
       routeBasePath: 'community',
     },
     ...buildPlugins,
