+++ config/jargon.js
