+++ config/svgo.config.js
