+++ tests/screenshot.spec.ts
@@ -5,8 +5,8 @@ import { extractSitemapPathnames, pathnameToArgosName } from './utils';
 
 // Constants
 const siteUrl = 'http://localhost:3000';
-const sitemapPath = './build/sitemap.xml';
-const stylesheetPath = './screenshot.css';
+const sitemapPath = require.resolve('../build/sitemap.xml');
+const stylesheetPath = require.resolve('./screenshot.css');
 const stylesheet = fs.readFileSync(stylesheetPath).toString();
 
 // Wait for hydration, requires Docusaurus v2.4.3+
